Consolidate freezer routines
Consolidate freezer routines used by the ActivityManagerService into a
new, single Freezer class.  ActivityManagerService.Injector provides a
Freezer to ActivityManagerService; this instance is used everywhere in
the ActivityManager code.

There is no change to behavior if the default Freezer instance (as
created by the default Injector) is used.  However, the injector can
supply a mocked Freezer (or a Freezer subclass) for testing.

Test: atest
 * FrameworksServicesTests:com.android.server.am
 * FrameworksMockingServicesTests:com.android.server.am

Flag: EXEMPT refactor
Bug: 326315985
Change-Id: I9da0b3dc0f52b1e1573e56b3cbfaebc74f830f42