Possible fix and additional logging for b/337742708
Logs indicate that the keyguard status bar is switching to visible
at startup due to the presence of the message "Updating keyguard
status bar state to visible" early in the shade log. This can only
happen in NPVC.StatusBarStateListener.onStateChanged, meaning the
status bar is possibly being shown on AOD due to bad state during
initialization of NPVC. This change adds a boolean that is set
during onViewAttachedToWindow to bypass the code that could
theoretically show the status bar on AOD during init.

Test: verified that the status bar still shows on lockscreen
Bug: 337742708
Flag: EXEMPT bug fix
Change-Id: I8afe59abcfd82c27c37af9aa2beb7cde8bb1a91a