audio: Fix permission update race
There are some subtle races on iterating over a concurrent queue and
aligning the task queue to the scheduler ordering.

Move to an explicitly lock a regular list, which is fine since the lock
is sparsely contended, has extremely tight critical sections, and is
easier to reason about.

Test: Manual + Cts
Flag: com.android.media.audio.audioserver_permissions
Bug: 362409920
Change-Id: I63c473d64a9b4b687c46b11dec882ea295ce6d36