Add ElementKey.placeAllCopies
This CL introduce ElementKey.placeAllCopies that makes an element place
all its copies during transitions, even when it is shared. This can be
useful for container elements that can have different contents in
different scenes/overlays (like in the dual shade case) but for which
the size and position should still be shared.

Bug: 353679003
Test: atest ElementTest
Flag: com.android.systemui.scene_container
Change-Id: I17675ea0abd3972a7f5c54bed63414f09b2c9b3a