Test display event delivery to cached processes.
Verify that a cached process does not receive display notifications.

This CL is a port of the CTS test DisplayEventTest into the unit-test
infrastructure.  A follow-on CL will add tests to verify that a frozen
process does not receive display notfications.  The new tests must use
platform APIs to interact with the freezer, but platform APIs are not
permitted in CTS tests.

In addition to the atest:
 * verified that this test runs in the DisplayServicesTests group.
 * verified that altering the DisplayManagerService UidListener code
   will trigger failures in this test.
 * verified that altering the DisplayManagerService isUidCached code
   will trigger failures in this test.

Flag: EXEMPT test-only
Test: atest
 * DisplayServicesTests
Bug: 326315985
Change-Id: Iaa7516a31a9d30e0601e8ce510f7ed89f82bbc91