Reset screenshot timeout immediately
We were only resetting the timeout after the animation finished, meaning
that if a second screenshot was invoked toward the end of the timeout
period, the old timeout could fire (immediately dismissing the
screenshot UI) even though there was a new screenshot visible.

This switches to resetting the timeout as soon as we get the new
screenshot.

Bug: 360043261
Fix: 360043261
Test: ensure that the screenshot UI doesn't time out if there's a second
screenshot, even if it's invoked at the end of the timeout period
Flag: EXEMPT trivial change

Change-Id: I5aee655880fcd0693ef45b18ae09be9c4c4d0d52