Remove App Header as an inset source when in full immersive
The App Header visibility is transient when in full immersive mode. It
follows the status bar visibility which is controlled by the user on
swipes on the system bar area. This means the app should not see the App
Header as an inset since it doesn't need to ajust to it. This change
adds a param to skip inset updates when in full immersive mode.

Flag: com.android.window.flags.enable_fully_immersive_in_desktop
Bug: 369444183
Bug: 369444147
Bug: 369443876
Test: atest WMShellUnitTests
Change-Id: I27daf861d37c3817c3ed74e42950ab28692492a9