Flag for cache invalidation when users are added or removed.
There is a list of methods which could be cached under the same key because they are invalidated at the same time. This includes things which are related to adding or removing UserData from mUserList in UserManagerService.

The cache should not be invalidated only when a user is removed, because if a client makes a binder call on a non-existing user then the cache would store the result. Then when a user with this exact id is created and the cache would not be invalidated then it would keep invalid data.

The User Serial number is one of those caches. The caches which should have a shared cache api name (different debug name if possible), so it could be invalidated at the same time and efficiently allocate a limited number of cache keys.

Change-Id: Iba99454501b65f1a5f2d389f3e1a2966e6776a15
Flag: android.multiuser.invalidate_cache_on_users_changed_read_only
Test: N/A
Bug:372383485