Fix showing the IME on another display, when requested from a virtual display.
A virtual display might initially have no registered (IME) insets source provider. In this case, a show request (from an app on the virtual display) was not executed, as it happened before the input target is set (only thereafter the source provider is set).
If the IME is on another display, we're setting the requested IME state on the remote target after updating the IME input target.

Test: atest VirtualDeviceImeTest
Bug: 372218080
Flag: android.view.inputmethod.refactor_insets_controller
Change-Id: Ieacd52fecd0acf82b1ee0cc9be80305d526bc77c