[events] move SystemAnimationState to enum, expose flow
This change is mostly mechanical; move from the @IntDef to using a bona
fide `enum class`. This makes `when` statements nicer when we attach
listeners to this newly-exposed flow.

The `animationState` flow now is an API that clients can use to track
the current animation state. This will enable us to attach view-binder
style animations to the state without needing to use the older callback
mechanism that collects `AnimatorSet`s to play them together.

Test: SystemStatusAnimationSchedulerImplTest
Test: AuthContainerViewTest
Flag: EXEMPT refactor
Bug: 364360986
Change-Id: I478d974f8525f6cc401db7c4390e0fad6b2d5090